+++ changelog.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
-## Unreleased [0.0.2]
+# Unreleased [0.0.2]
+
+## Added
+
+- Correct or add profile correctly visbale at admin side.
+
+## Fixed 
+
+- [#2](https://github.com/softtechVI/Examiify-frontend/issues/1) login is succes alert show
+- [#2](https://github.com/softtechVI/Examiify-frontend/issues/2) text is already show
+- [#3](https://github.com/softtechVI/Examiify-frontend/issues/3) Create an Dashboard Component
 
 
 - [#21](https://github.com/softtechVI/Examify/issues/21) all changes merge till this.